fix(BannerRibbon): titleWidthHint prop should force float the fit rest
Currently, the titleWidthHint only informs the width of the title section. Now, it provides a hint
to the float content to the right which forces it to take up the remaining space. This behavior can
be opted out of by wrapping your floatRightOfTitle section with another element but this makes the
most sense as a default.

Fixes 52