fix(Elements/VerticalScrollableContent, Grid, BannerRibbon): Lock down scrolling to vertical only on
If someone needs horizontal scrolling, they can set the css but they need to be aware that grids
will always try to take 100vw which includes the width of the scrollbar. This causes a persistant
horizontal scrollbar if a grid is within the verticalscrollablecontent.

Disabling horizontal scroll on verticalscrollablecontent, as name implies.