feat(Mobile): Add Mobile component
The Mobile component is a shortcut to react-media with a little helper prop that immediately works
with our predefined breakpoints. All you need to do is provide a valid max prop (based on available
breakpoints, base, xs, s, m, l, and xl) and it will deal with making the check.